pyproject.toml
[tool.poetry]
name = "sms-110-linebot"
version = "0.1.0"
description = ""
authors = ["FOBshippingpoint <aulaulql4ql4xj6@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
line-bot-sdk = "^2.2.1"
Flask = "^2.1.2"
bs4 = "^0.0.1"
pyimgur = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
